Refactor highlighting into custom StyledDocument
This fixes the issue where inline images would get mangled, and partly
decouples GUI from the document model.

RES-660
